fix: Forwarding files not working because docs not forwarded in state
When we want to share multiple docs, they are forwarded to
ShareBottomSheet thanks to react-router-dom navigate state. But this
state does not support every data types (see
https://html.spec.whatwg.org/multipage/structured-data.html#structuredserializeinternal).
Notably it does not support functions and symbols.

We added recently the `bills` doctypes and the relationship with
`files` and it was resulting in a "DataCloneError" DOMException
because it is possible to have functions in `bills`.