feat: Add prefix `paper` to all routes
With the merge of cozy-mespapiers-lib we had this prefix.
We wanted to take advantage of this work to simplify the url,
but several apps/libs use redirection urls including this prefix.

In the end, it doesn't change much...
The fix is faster if you put it back on the app.