refactor(ShareBottomSheet): Use search params instead state
...to transmit files.
State is much less resilient on a url that can be reloaded.
It is preferable to use the ids in search params
and query if necessary.

Modification of the `buildFileQueryById` query is necessary
since the `buildFileQueryByIds` query has been added,
`buildFileQueryById` always returns a file in an array
(even with `singleDocData: true`)
IF it is called second in the app lifecycle.