Merge `share` route with `forward` route
In the end, the `share` route was only useful
for displaying a BottomSheet.
Since we don't want to return to the history of the latter,
it's better to manage it with a local state of the parent.

This new parent makes it possible to combine the logic
of the `ForwardModalByRoute` & `ShareBottomSheetByRoute`
components and return a BottomSheet or
modal according to simple conditions.