feat: Change the behavior of PagePickerModal
Visuals and behavior change.

Selecting the 2 sides no longer means that the file is
to be separated, but rather that a third option must
be checked to explicitly state that the file is to be separated.

To be fairer to the native API, I've also renamed
the `value` attribute to `name`.