feat: Ensure the app settings document exists when the app is loaded
In the case of first-time users, some functions need
this document to work properly, so we need to
make sure it exists.

We've had this regression since the removal of the
onboarding screen, which took care of this among other things.
(cozy/cozy-libs@a5e7d28)

To clarify, each time the app is loaded (new or reloaded),
the `metadataMigration` job is created if never
created during the day.
This job ends up updating the document
(creating it if it doesn't exist),
but as the jobs are not synchronous,
the problem of the missing document exists.