fix: Open country selection for FR papers
Two documents have a slightly different route between
their FR and foreign versions: the driving license and
the national identity card.

The creation of one of these FR papers does not offer
a country selection step, which poses a problem if you
wish to edit the country after the paper has been created.

The best approach is to load the paperDefinition of
the foreign version to open the country selection step.